client/game/hud: Fixed the stats widget
The way it was done it was causing a lot of CPU issues
